+++ ChangeLog.md
@@ -6,4 +6,9 @@
 
 * help button added
 
+* improved search
+  - diacritics ignored
+  - separate search in song number (exact), title, author or source (all subset)
+
+
 # 0.4.1